fix: handlers -> handler
