--- app/process/handler/consensus-message/consensus-message.go
@@ -1,3 +1,0 @@
-package consensusmessage
-
-// TODO: Consensus message event handler
